# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenG2P Approval Workflow Engine",
    "description": "Generic, configurable multi-stage approval workflow engine for OpenG2P modules. Caller services post artifacts; AWE resolves stages and approvers, then notifies the caller via signed webhook callbacks when state changes.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/awe/policies": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List policies (newest version of each policy_key)",
        "operationId": "list_policies_v1_awe_policies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicyOut"
                  },
                  "type": "array",
                  "title": "Response List Policies V1 Awe Policies Get"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks `AWE_VIEWER` or `AWE_ADMIN` role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_VIEWER or AWE_ADMIN role"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Create the first draft of a new policy",
        "operationId": "create_policy_v1_awe_policies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks the `AWE_ADMIN` role required for this mutation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_ADMIN role"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Policy version conflict \u2014 typically a duplicate `policy_key` on create, or attempting to edit/activate a version in the wrong status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-002",
                      "message": "policy_key 'registry.cr.v1' already exists (latest v3)"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/awe/policies/{policy_key}/versions": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "List all versions of a policy_key",
        "operationId": "list_versions_v1_awe_policies__policy_key__versions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "policy_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicyVersionOut"
                  },
                  "title": "Response List Versions V1 Awe Policies  Policy Key  Versions Get"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks `AWE_VIEWER` or `AWE_ADMIN` role.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_VIEWER or AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No matching policy / policy version exists.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-001",
                      "message": "Policy 'registry.cr.v1' not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/policies/{policy_key}/versions/{version}": {
      "get": {
        "tags": [
          "policies"
        ],
        "summary": "Fetch a specific policy version with stages and rules",
        "operationId": "get_version_v1_awe_policies__policy_key__versions__version__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "policy_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Key"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks `AWE_VIEWER` or `AWE_ADMIN` role.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_VIEWER or AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No matching policy / policy version exists.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-001",
                      "message": "Policy 'registry.cr.v1' not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "policies"
        ],
        "summary": "Edit a draft version in place (drafts only)",
        "operationId": "edit_draft_v1_awe_policies__policy_key__versions__version__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "policy_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Policy Key"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks the `AWE_ADMIN` role required for this mutation.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Policy body violates a structural rule (e.g. URL/body key mismatch).",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-010",
                      "message": "policy_key in body must match URL"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No matching policy / policy version exists.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-001",
                      "message": "Policy 'registry.cr.v1' not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource is in a state that disallows this transition. Examples: request is already in a terminal state (`approved`/`rejected`/`cancelled`); task is not `open`/`claimed`; policy version is not `draft` for editing; delivery has already succeeded.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-007",
                      "message": "Request is in terminal state 'approved' \u2014 cannot cancel"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": "openg2p.awe",
            "examples": [
              "openg2p.awe"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0",
            "examples": [
              "1.0"
            ]
          },
          "responsetime": {
            "type": "string",
            "title": "Responsetime",
            "examples": [
              "2026-04-23T10:00:00.000Z"
            ]
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "responsetime",
          "errors"
        ],
        "title": "ErrorResponse"
      },
...
      "PolicyCreate": {
        "properties": {
          "policy_key": {
            "type": "string",
            "title": "Policy Key",
            "examples": [
              "registry.change_request.v1"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Registry CR approval"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type",
            "examples": [
              "registry.change_request"
            ]
          },
          "forbid_self_approval": {
            "type": "boolean",
            "title": "Forbid Self Approval",
            "description": "Filter the request's requester out of every stage's approver list.",
            "default": false
          },
          "forbid_repeat_approvers": {
            "type": "boolean",
            "title": "Forbid Repeat Approvers",
            "description": "Filter users who approved an earlier stage out of later stages.",
            "default": false
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageIn"
            },
            "type": "array",
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "policy_key",
          "name",
          "artifact_type"
        ],
        "title": "PolicyCreate"
      },
      "PolicyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "policy_key": {
            "type": "string",
            "title": "Policy Key"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type"
          },
          "forbid_self_approval": {
            "type": "boolean",
            "title": "Forbid Self Approval",
            "default": false
          },
          "forbid_repeat_approvers": {
            "type": "boolean",
            "title": "Forbid Repeat Approvers",
            "default": false
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageOut"
            },
            "type": "array",
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "policy_key",
          "version",
          "name",
          "status",
          "artifact_type",
          "created_at",
          "updated_at"
        ],
        "title": "PolicyOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}